ppc32: use double for unw_tdep_fpreg_t, not long double
ppc32 floating-point registers are 64-bit IEEE 754 values, stored as
double (8 bytes) in fpregset_t::fpregs[].  access_fpreg() in Ginit.c
retrieves their address and casts it to unw_fpreg_t * for load/store,
but unw_tdep_fpreg_t was long double (16 bytes, 16-byte alignment).

This caused two bugs caught by UBSAN:
- misaligned store/load: double* is 8-byte aligned, long double* needs 16
- size mismatch: the 16-byte store would also overwrite the adjacent
  element in the fpregs[] array

dwarf_getfp/dwarf_putfp already transfer only one unw_word_t (8 bytes)
regardless of the declared fpreg type, so long double was never correct
here.  ppc64 uses double for unw_tdep_fpreg_t; align ppc32 with that.

+++ include/libunwind-ppc32.h
@@ -77,7 +77,7 @@ typedef int64_t unw_sword_t;
 # define UNW_WORD_MAX UINT64_MAX
 #endif
 
-typedef long double unw_tdep_fpreg_t;
+typedef double unw_tdep_fpreg_t;
 
 typedef enum
   {
